backend/requirements.txt
fastapi
uvicorn
librosa
soundfile
ffmpeg-python
xgboost
speechbrain
numpy
scikit-learn
joblib
torch
python-multipart
